refactor(evaluation): move pure helpers out of evaluator.py into helpers module
Extracts `merge_blueprint_into_config`, `resolve_dataset_items_generator`
(+ its private companion `_calculate_total_items`), and the
`EVALUATION_STREAM_DATASET_BATCH_SIZE` constant from evaluator.py into
`opik.evaluation.helpers`. The existing `_compute_experiment_scores` and
`_try_notifying_about_experiment_completion` stay in evaluator.py.

Tests for the moved functions now live in `test_helpers.py` — the pure
ones (`_calculate_total_items`, `resolve_project_name`) are fully
mock-free, exercised via raw values or `SimpleNamespace` stand-ins.
